pyproject.toml
[project]
name = "text-analysis-app"
version = "0.2.0"
description = "CLI app for analyzing texts form .xlsx using SpaCy"
authors = [{ name = "Sara Barac" }]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "spacy>=3.7,<4.0",
  "typer>=0.12",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "pandas>=2.0",
  "openpyxl>=3.1.0", 
  "numpy>=1.26",
  "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl",
  "en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl"
  ]



[project.scripts]
textan = "text_analysis_app.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["text_analysis_app*"]